fixes some tests
