+++ dlt/destinations/impl/sqlalchemy/sqlalchemy_job_client.py
@@ -49,7 +49,6 @@ class SqlalchemyJobClient(SqlJobClientWithStagingDataset):
             config.normalize_staging_dataset_name(schema),
             config.credentials,
             capabilities,
-            engine_args=config.engine_args,
         )
 
         self.schema = schema
